% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstras.R
\name{dijkstra}
\alias{dijkstra}
\title{Distance Computation Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe with three numeric columns viz. "v1", "v2", and "w", representing each node, its adjacent node, and the distance between them, respectively}

\item{init_node}{The starting node}
}
\value{
Shortest distance from \strong{\code{init_node}} to all other nodes in the \strong{\code{graph}}.
}
\description{
Given a connected graph and starting node, find the shortest distance from the starting node to every other node in the graph using Dijkstra's algorithm
}
\references{


\href{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}{Wiki page on what a graph is and more}
}
