% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclid.djikstras.R
\docType{package}
\name{euclid.dijkstras}
\alias{euclid.dijkstras}
\title{euclid.dijkstras}
\description{
A package for two functions: one to compute the greatest common divisor of two numbers using Euclid's algorithm, another to compute the shortest distance between vertices in a graph using Dijkstra's algorithm
}
\details{
The two functions in the package are:
\strong{euclidean} and \strong{dijkstra}
}
\section{More details available here}{

\code{\link{euclidean}}
\code{\link{dijkstra}}
}

