\name{Euclid_Dijkstras-package}
\alias{Euclid_Dijkstras-package}
\alias{Euclid_Dijkstras}
\docType{package}
\title{
\packageTitle{Euclid_Dijkstras}
}
\description{
\packageDescription{Euclid_Dijkstras}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclid_Dijkstras}
\packageIndices{Euclid_Dijkstras}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Euclid_Dijkstras}

Maintainer: \packageMaintainer{Euclid_Dijkstras}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
