% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Get GCD of two numbers}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
Greatest common divisor of \strong{\code{x}} and \strong{\code{y}}.
}
\description{
Implements Euclid's algorithm to compute the greatest common divisor of two numbers
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{Euclidean Algorithm}
}
