% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstras.R
\name{dijkstra}
\alias{dijkstra}
\title{Distance Computation Algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A dataframe with three numeric columns viz. "v1", "v2", and "w", representing each node, its adjacent node, and the distance between them, respectively}

\item{init_node}{The starting node}
}
\value{
Shortest distance from \strong{\code{init_node}} to all other nodes in the \strong{\code{graph}}.
}
\description{
Given a connected graph and starting node, find the shortest distance from the starting node to every other node in the graph using Dijkstra's algorithm
}
\examples{
wiki_graph <- data.frame(v1=c(1,1,1,2,2,2,3,3,3,3,4,4,4,5,5,6,6,6),v2=c(2,3,6,1,3,4,1,2,4,6,2,3,5,4,6,1,3,5),w=c(7,9,14,7,10,15,9,10,11,2,15,11,6,6,9,14,2,9))
dijkstra(wiki_graph,1)
dijkstra(wiki_graph,3)

}
\references{
\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Dijkstra's Algorithm}

\href{https://en.wikipedia.org/wiki/Graph_(discrete_mathematics)}{Wiki page on what a graph is and more}
}
